Week 6 Live Class done-> Added authentication & authorization

+++ src/main/java/com/AirBndProject/service/RoomServiceImpl.java
@@ -3,13 +3,16 @@ package com.AirBndProject.service;
 import com.AirBndProject.dto.RoomDto;
 import com.AirBndProject.entities.Hotel;
 import com.AirBndProject.entities.Room;
+import com.AirBndProject.entities.User;
 import com.AirBndProject.exceptions.ResourceNotFoundException;
+import com.AirBndProject.exceptions.UnAuthorizedException;
 import com.AirBndProject.repository.HotelRepository;
 import com.AirBndProject.repository.RoomRepository;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.modelmapper.ModelMapper;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -35,6 +38,11 @@ public class RoomServiceImpl implements RoomService{
                 .orElseThrow(()-> new ResourceNotFoundException(("Hotel not found with Id -> " + hotelId)));
 
         log.info("Hotel found -> {}",hotel.getId());
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!user.equals(hotel.getOwner()))
+        {
+            throw new UnAuthorizedException("This user don not own this hotel with id ->"+hotelId);
+        }
 
         Room room = modelMapper.map(rootDto,Room.class);
         room.setHotel(hotel);
@@ -59,6 +67,12 @@ public class RoomServiceImpl implements RoomService{
         Hotel hotel = hotelRepository.findById(hotelId)
                 .orElseThrow(() -> new ResourceNotFoundException("Hotel not found with Id -> " + hotelId));
 
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!user.equals(hotel.getOwner()))
+        {
+            throw new UnAuthorizedException("This user don not own this hotel with id ->"+hotelId);
+        }
+
         return hotel.getRooms()
                 .stream()
                 .map((element) -> modelMapper
@@ -82,6 +96,11 @@ public class RoomServiceImpl implements RoomService{
     {
         Room room = roomRepository.findById(roomId)
                 .orElseThrow(() -> new ResourceNotFoundException("Room not found with Id -> " + roomId));
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!user.equals(room.getHotel().getOwner()))
+        {
+            throw new UnAuthorizedException("This user don not own this room with id ->"+room.getHotel().getCity());
+        }
 
         inventoryService.deleteAllinventories(room);
         roomRepository.deleteById(roomId);
